# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BookIt API
  version: 1.0.0
  description: REST API сервиса онлайн-записи BookIt.
servers:
  - url: http://localhost:3000/api
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [user, admin]
        timezone:
          type: string
        created_at:
          type: string
          format: date-time
...
    TimeSlot:
      type: object
      properties:
        id:
          type: integer
        service_id:
          type: integer
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
        slot_time:
          type: string
          example: "14:00"
        max_bookings:
          type: integer
        is_active:
          type: integer
        created_at:
          type: string
          format: date-time
...
    ProfileStats:
      type: object
      properties:
        active:
          type: integer
        cancelled:
          type: integer
        total:
          type: integer
        total_spent:
          type: integer
    ProfileResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        stats:
          $ref: "#/components/schemas/ProfileStats"
      required:
        - user
        - stats
...
paths:
...
  /profile:
    get:
      tags: [Profile]
      summary: Профиль пользователя
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Профиль и статистика
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
    put:
      tags: [Profile]
      summary: Обновить профиль
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                timezone:
                  type: string
      responses:
        "200":
          description: Профиль обновлён
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
                  token:
                    type: string
        "400":
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profile/password:
    put:
      tags: [Profile]
      summary: Сменить пароль
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
      responses:
        "200":
          description: Пароль изменён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /slots:
    post:
      tags: [Slots]
      summary: Добавить слот в расписание
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_id, day_of_week, slot_time]
              properties:
                service_id:
                  type: integer
                day_of_week:
                  type: integer
                  minimum: 0
                  maximum: 6
                slot_time:
                  type: string
                  example: "09:00"
                max_bookings:
                  type: integer
                  example: 2
      responses:
        "201":
          description: Слот создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  slot:
                    $ref: "#/components/schemas/TimeSlot"
  /slots/{id}:
    delete:
      tags: [Slots]
      summary: Деактивировать слот
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Слот деактивирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /slots/service/{serviceId}:
    get:
      tags: [Slots]
      summary: Расписание услуги
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - in: path
          name: serviceId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Расписание
          content:
            application/json:
              schema:
                type: object
                properties:
                  slots:
                    type: array
                    items:
                      $ref: "#/components/schemas/TimeSlot"